backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# AI & ML
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
tiktoken==0.5.2

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
opencv-python==4.9.0.80

# Task Queue
celery==5.3.6
redis==5.0.1

# Notifications
twilio==8.11.1
sendgrid==6.11.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
